requirements.txt
torch
torchvision
transformers
pyarrow
pillow
pyyaml
numpy
scikit-learn
